Created my own implementation of List type
Just a simple List
