+++ GameEngine/GameEngine/GameEngine.vcxproj
@@ -150,13 +150,15 @@
     <ClCompile Include="src\EventSystem\System.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="src\EventSystem\Queue.h" />
+    <ClInclude Include="src\Utils\SimpleList.h" />
+    <ClInclude Include="src\Utils\Queue.h" />
     <ClInclude Include="src\EventSystem\Message.h" />
     <ClInclude Include="src\EventSystem\MessageBus.h" />
     <ClInclude Include="src\EventSystem\System.h" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="src\EventSystem\Queue.tpp" />
+    <None Include="src\Utils\SimpleList.tpp" />
+    <None Include="src\Utils\Queue.tpp" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
