+++ GameEngine/GameEngine/src/Utils/SimpleList.h
@@ -1,0 +1,19 @@
+#pragma once
+
+template <typename T>
+class SimpleList
+{
+public:
+	SimpleList();
+
+	void add(T* value);
+	T get(int index);
+private:
+	int maxLength;
+	int currentIndex;
+	T* list;
+
+	T* resize(int size);
+};
+
+#include "SimpleList.tpp"
